Move compatibility passes to separate package.
