+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/layers/EnhancedBaseCreator.scala
@@ -4,7 +4,8 @@ import io.shiftleft.SerializedCpg
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.Languages
 import io.shiftleft.passes.CpgPass
-import io.shiftleft.semanticcpg.passes.bindingtablecompat.BindingTableCompat
+import io.shiftleft.semanticcpg.passes.compat.bindingtablecompat.BindingTableCompat
+import io.shiftleft.semanticcpg.passes.compat.callnamecompat.CallNameFixup
 import io.shiftleft.semanticcpg.passes.containsedges.ContainsEdgePass
 import io.shiftleft.semanticcpg.passes.languagespecific.fuzzyc.{MethodStubCreator, TypeDeclStubCreator}
 import io.shiftleft.semanticcpg.passes.linking.calllinker.CallLinker
@@ -13,7 +14,7 @@ import io.shiftleft.semanticcpg.passes.linking.linker.Linker
 import io.shiftleft.semanticcpg.passes.linking.memberaccesslinker.MemberAccessLinker
 import io.shiftleft.semanticcpg.passes.methoddecorations.MethodDecoratorPass
 import io.shiftleft.semanticcpg.passes.methodexternaldecorator.MethodExternalDecoratorPass
-import io.shiftleft.semanticcpg.passes.methodinstcompat.MethodInstCompat
+import io.shiftleft.semanticcpg.passes.compat.methodinstcompat.MethodInstCompat
 import io.shiftleft.semanticcpg.passes.namespacecreator.NamespaceCreator
 import io.shiftleft.semanticcpg.passes.receiveredges.ReceiverEdgePass
 
@@ -41,6 +42,21 @@ class EnhancedBaseCreator(cpg: Cpg, language: String, serializedCpg: SerializedC
           new MethodInstCompat(cpg),
           new TypeDeclStubCreator(cpg),
           new MethodStubCreator(cpg),
+          new MethodDecoratorPass(cpg),
+          new CapturingLinker(cpg),
+          new Linker(cpg),
+          new BindingTableCompat(cpg),
+          new CallLinker(cpg),
+          new MemberAccessLinker(cpg),
+          new MethodExternalDecoratorPass(cpg),
+          new ContainsEdgePass(cpg),
+          new NamespaceCreator(cpg),
+        )
+      case Languages.CSHARP =>
+        List(
+          new MethodInstCompat(cpg),
+          new CallNameFixup(cpg),
+          new ReceiverEdgePass(cpg),
           new MethodDecoratorPass(cpg),
           new CapturingLinker(cpg),
           new Linker(cpg),
